Refactor sidebar functions to sidebar module
 - Move trends_list to constants module
